proofreading corrections chap 7-19 (#17)
I fixed all proofreading errors found by Adrienne, but at some point I need to go through and fix the formatting for narrator comments, either putting them all in brackets (as a few already are), or putting Healy's words in quotes. The entire book weaves in and out of narration, and none of it is formatted well.